lumi/lumi_backend/emotion_backend/requirements.txt
tqdm>=4.66.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.4.0
opencv-python>=4.10.0.84
tensorboard>=2.17.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
typing-extensions>=4.12.0
